Fixed dagger bug with images fragment & product viewModel
